requirements.txt
streamlit
pandas
numpy
pdfplumber
python-docx
openpyxl
